Added very crude client-server implementation
The server interfaces with the sqlite backend and provides data to
multiple clients. The data is transmitted over the network using tcp and
very simple message format using json. The code is still really messy
(especially the serialization code). The next efforts will be spent
making this code presentable.

+++ gui/dialogs/editreservation.cpp
@@ -23,7 +23,7 @@ namespace gui
       {
         Resolution resolution;
         resolution.merged = base;
-        for (int i = 0; i < base.size(); ++i)
+        for (int i = 0; i < static_cast<int>(base.size()); ++i)
         {
           bool oursChanged = base[i] != ours[i];
           bool theirsChanged = base[i] != theirs[i];
